Refactor eth.go
