+++ node/impl/full/eth_trace.go
@@ -2,48 +2,21 @@ package full
 
 import (
 	"bytes"
-	"context"
 	"encoding/binary"
 	"encoding/hex"
 	"fmt"
 	"io"
 
 	"github.com/ipfs/go-cid"
-	"go.uber.org/fx"
-	"golang.org/x/xerrors"
 
 	"github.com/filecoin-project/go-state-types/abi"
 	"github.com/filecoin-project/go-state-types/builtin"
 	"github.com/filecoin-project/go-state-types/exitcode"
 
-	"github.com/filecoin-project/lotus/api"
 	builtinactors "github.com/filecoin-project/lotus/chain/actors/builtin"
-	"github.com/filecoin-project/lotus/chain/stmgr"
-	"github.com/filecoin-project/lotus/chain/store"
 	"github.com/filecoin-project/lotus/chain/types"
 	"github.com/filecoin-project/lotus/chain/types/ethtypes"
 )
-
-type EthTraceAPI interface {
-	TraceBlock(ctx context.Context, blkNum string) (interface{}, error)
-	TraceReplayBlockTransactions(ctx context.Context, blkNum string, traceTypes []string) (interface{}, error)
-}
-
-var (
-	_ EthTraceAPI = *new(api.FullNode)
-)
-
-type EthTrace struct {
-	fx.In
-
-	Chain        *store.ChainStore
-	StateManager *stmgr.StateManager
-
-	ChainAPI
-	EthModuleAPI
-}
-
-var _ EthTraceAPI = (*EthTrace)(nil)
 
 type Trace struct {
 	Action       Action `json:"action"`
@@ -91,168 +64,6 @@ type Action struct {
 type Result struct {
 	GasUsed ethtypes.EthUint64 `json:"gasUsed"`
 	Output  string             `json:"output"`
-}
-
-func (e *EthTrace) TraceBlock(ctx context.Context, blkNum string) (interface{}, error) {
-	ts, err := e.getTipsetByBlockNr(ctx, blkNum, false)
-	if err != nil {
-		return nil, err
-	}
-
-	_, trace, err := e.StateManager.ExecutionTrace(ctx, ts)
-	if err != nil {
-		return nil, xerrors.Errorf("failed to compute base state: %w", err)
-	}
-
-	tsParent, err := e.ChainAPI.ChainGetTipSetByHeight(ctx, ts.Height()+1, e.Chain.GetHeaviestTipSet().Key())
-	if err != nil {
-		return nil, fmt.Errorf("cannot get tipset at height: %v", ts.Height()+1)
-	}
-
-	msgs, err := e.ChainGetParentMessages(ctx, tsParent.Blocks()[0].Cid())
-	if err != nil {
-		return nil, err
-	}
-
-	cid, err := ts.Key().Cid()
-	if err != nil {
-		return nil, err
-	}
-
-	blkHash, err := ethtypes.EthHashFromCid(cid)
-	if err != nil {
-		return nil, err
-	}
-
-	allTraces := make([]*TraceBlock, 0, len(trace))
-	for _, ir := range trace {
-		// ignore messages from f00
-		if ir.Msg.From.String() == "f00" {
-			continue
-		}
-
-		idx := -1
-		for msgIdx, msg := range msgs {
-			if ir.Msg.From == msg.Message.From {
-				idx = msgIdx
-				break
-			}
-		}
-		if idx == -1 {
-			log.Warnf("cannot resolve message index for cid: %s", ir.MsgCid)
-			continue
-		}
-
-		txHash, err := e.EthGetTransactionHashByCid(ctx, ir.MsgCid)
-		if err != nil {
-			return nil, err
-		}
-		if txHash == nil {
-			log.Warnf("cannot find transaction hash for cid %s", ir.MsgCid)
-			continue
-		}
-
-		traces := []*Trace{}
-		err = buildTraces(&traces, nil, []int{}, ir.ExecutionTrace, int64(ts.Height()))
-		if err != nil {
-			return nil, xerrors.Errorf("failed when building traces: %w", err)
-		}
-
-		traceBlocks := make([]*TraceBlock, 0, len(trace))
-		for _, trace := range traces {
-			traceBlocks = append(traceBlocks, &TraceBlock{
-				Trace:               trace,
-				BlockHash:           blkHash,
-				BlockNumber:         int64(ts.Height()),
-				TransactionHash:     *txHash,
-				TransactionPosition: idx,
-			})
-		}
-
-		allTraces = append(allTraces, traceBlocks...)
-	}
-
-	return allTraces, nil
-}
-
-func (e *EthTrace) TraceReplayBlockTransactions(ctx context.Context, blkNum string, traceTypes []string) (interface{}, error) {
-	if len(traceTypes) != 1 || traceTypes[0] != "trace" {
-		return nil, fmt.Errorf("only 'trace' is supported")
-	}
-
-	ts, err := e.getTipsetByBlockNr(ctx, blkNum, false)
-	if err != nil {
-		return nil, err
-	}
-
-	_, trace, err := e.StateManager.ExecutionTrace(ctx, ts)
-	if err != nil {
-		return nil, xerrors.Errorf("failed when calling ExecutionTrace: %w", err)
-	}
-
-	allTraces := make([]*TraceReplayBlockTransaction, 0, len(trace))
-	for _, ir := range trace {
-		// ignore messages from f00
-		if ir.Msg.From.String() == "f00" {
-			continue
-		}
-
-		txHash, err := e.EthGetTransactionHashByCid(ctx, ir.MsgCid)
-		if err != nil {
-			return nil, err
-		}
-		if txHash == nil {
-			log.Warnf("cannot find transaction hash for cid %s", ir.MsgCid)
-			continue
-		}
-
-		t := TraceReplayBlockTransaction{
-			Output:          hex.EncodeToString(ir.MsgRct.Return),
-			TransactionHash: *txHash,
-			StateDiff:       nil,
-			VmTrace:         nil,
-		}
-
-		err = buildTraces(&t.Trace, nil, []int{}, ir.ExecutionTrace, int64(ts.Height()))
-		if err != nil {
-			return nil, xerrors.Errorf("failed when building traces: %w", err)
-		}
-
-		allTraces = append(allTraces, &t)
-	}
-
-	return allTraces, nil
-}
-
-func writePadded[T any](w io.Writer, data T, size int) error {
-	tmp := &bytes.Buffer{}
-
-	// first write data to tmp buffer to get the size
-	err := binary.Write(tmp, binary.BigEndian, data)
-	if err != nil {
-		return fmt.Errorf("writePadded: failed writing tmp data to buffer: %w", err)
-	}
-
-	if tmp.Len() > size {
-		return fmt.Errorf("writePadded: data is larger than size")
-	}
-
-	// write tailing zeros to pad up to size
-	cnt := size - tmp.Len()
-	for i := 0; i < cnt; i++ {
-		err = binary.Write(w, binary.BigEndian, uint8(0))
-		if err != nil {
-			return fmt.Errorf("writePadded: failed writing tailing zeros to buffer: %w", err)
-		}
-	}
-
-	// finally write the actual value
-	err = binary.Write(w, binary.BigEndian, tmp.Bytes())
-	if err != nil {
-		return fmt.Errorf("writePadded: failed writing data to buffer: %w", err)
-	}
-
-	return nil
 }
 
 // buildTraces recursively builds the traces for a given ExecutionTrace by walking the subcalls
@@ -416,6 +227,37 @@ func buildTraces(traces *[]*Trace, parent *Trace, addr []int, et types.Execution
 	return nil
 }
 
+func writePadded[T any](w io.Writer, data T, size int) error {
+	tmp := &bytes.Buffer{}
+
+	// first write data to tmp buffer to get the size
+	err := binary.Write(tmp, binary.BigEndian, data)
+	if err != nil {
+		return fmt.Errorf("writePadded: failed writing tmp data to buffer: %w", err)
+	}
+
+	if tmp.Len() > size {
+		return fmt.Errorf("writePadded: data is larger than size")
+	}
+
+	// write tailing zeros to pad up to size
+	cnt := size - tmp.Len()
+	for i := 0; i < cnt; i++ {
+		err = binary.Write(w, binary.BigEndian, uint8(0))
+		if err != nil {
+			return fmt.Errorf("writePadded: failed writing tailing zeros to buffer: %w", err)
+		}
+	}
+
+	// finally write the actual value
+	err = binary.Write(w, binary.BigEndian, tmp.Bytes())
+	if err != nil {
+		return fmt.Errorf("writePadded: failed writing data to buffer: %w", err)
+	}
+
+	return nil
+}
+
 func handleFilecoinMethodInput(method abi.MethodNum, codec uint64, params []byte) ([]byte, error) {
 	NATIVE_METHOD_SELECTOR := []byte{0x86, 0x8e, 0x10, 0xc4}
 	EVM_WORD_SIZE := 32
@@ -476,40 +318,4 @@ func handleFilecoinMethodOutput(exitCode exitcode.ExitCode, codec uint64, data [
 	}
 
 	return w.Bytes(), nil
-}
-
-// TODO: refactor this to be shared code
-func (e *EthTrace) getTipsetByBlockNr(ctx context.Context, blkParam string, strict bool) (*types.TipSet, error) {
-	if blkParam == "earliest" {
-		return nil, fmt.Errorf("block param \"earliest\" is not supported")
-	}
-
-	head := e.Chain.GetHeaviestTipSet()
-	switch blkParam {
-	case "pending":
-		return head, nil
-	case "latest":
-		parent, err := e.Chain.GetTipSetFromKey(ctx, head.Parents())
-		if err != nil {
-			return nil, fmt.Errorf("cannot get parent tipset")
-		}
-		return parent, nil
-	default:
-		var num ethtypes.EthUint64
-		err := num.UnmarshalJSON([]byte(`"` + blkParam + `"`))
-		if err != nil {
-			return nil, fmt.Errorf("cannot parse block number: %v", err)
-		}
-		if abi.ChainEpoch(num) > head.Height()-1 {
-			return nil, fmt.Errorf("requested a future epoch (beyond 'latest')")
-		}
-		ts, err := e.ChainAPI.ChainGetTipSetByHeight(ctx, abi.ChainEpoch(num), head.Key())
-		if err != nil {
-			return nil, fmt.Errorf("cannot get tipset at height: %v", num)
-		}
-		if strict && ts.Height() != abi.ChainEpoch(num) {
-			return nil, ErrNullRound
-		}
-		return ts, nil
-	}
 }
